Move MavenDependencyResolver to its own artifact.
