+++ pluginapi/src/main/java/org/robolectric/internal/dependency/DependencyJar.java
