+++ pluginapi/src/main/java/org/robolectric/internal/dependency/DependencyResolver.java
