+++ settings.gradle
@@ -5,6 +5,8 @@ include ":sandbox"
 include ":junit"
 include ":utils"
 include ":utils:reflector"
+include ":pluginapi"
+include ":plugins:maven-dependency-resolver"
 include ":processor"
 include ":resources"
 include ":annotations"
